chore: places carbon tests where they belong
