+++ src/test/cassettes/shipment/regenerate_rates_with_carbon_offset.json
